pyproject.toml
[tool.poetry]
name = "boilerplate AML pipelines"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
azure-identity = "=1.4.1"
azure-cognitiveservices-vision-customvision = "^3.0.0"
azure-mgmt-resource = "=10.2.0"
azure-mgmt-network = "=16.0.0"
azure-mgmt-compute = "=17.0.0"
azure-storage-blob = "=12.5.0"
azureml-sdk = "=1.17.0"
numpy = "1.18.5"
pandas = "=1.1.4"
pytest-cov = "^2.10.1"
python = "=3.8.5"
python-dotenv = "0.14.0"
scikit-learn = "^0.23.2"
torch = {version = "1.6.0", python = "^3.8", platform = "linux"}

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
ipykernel = "^5.3.4"
jupyter = "^1.0.0"
papermill = "^2.2.0"
pylint = "^2.6.0"
pydub = "^0.24.1"
pytest = "^6.1.1"
yapf = "^0.30.0"
